+++ src/api/main/java/isen/projet_dp_api/bean/AuthConfig.java
@@ -1,4 +1,4 @@
-package isen.projet_dp_api.security;
+package isen.projet_dp_api.bean;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
